added calloc

+++ malloc.h
@@ -1,5 +1,6 @@
 #ifndef ADD_H_GUARD
 #define ADD_H_GUARD
-void *malloc(unsigned int size);
+void *malloc(size_t int size);
+//void *calloc(unsigned int size);
 void free(void* addr);
 #endif
